+++ tests/integration-tests/src/test/java/org/apache/activemq/artemis/tests/integration/paging/PagingWithFailoverAndCountersTest.java
@@ -357,7 +357,7 @@ public class PagingWithFailoverAndCountersTest extends ActiveMQTestBase {
       waitForServerToStart(server);
       Queue queue = server.locateQueue(SimpleString.toSimpleString("cons2"));
 
-      int messageCount = (int) getMessageCount(queue);
+      int messageCount = getMessageCount(queue);
 
       assertTrue(messageCount >= 0);
